Renaming "uwp" to "uwp_slug"
The UWP is the concatenation of the UWP slugs.  As I was writing code, I
was being lazy about defining the code as a full UWP.  Instead, I want
the variable name to be more descriptive.

I thought about saying its the uwp_code, but prefer slug as there's a
positional implication.